+++ src/rtlib/fb_data.h
@@ -18,7 +18,7 @@ extern FB_DATADESC *__fb_data_ptr;
 
 FBCALL void         fb_DataRestore      ( FB_DATADESC *labeladdr );
        void         fb_DataNext         ( void );
-FBCALL void         fb_DataReadStr      ( void *dst, int dst_size, int fillrem );
+FBCALL void         fb_DataReadStr      ( void *dst, ssize_t dst_size, int fillrem );
 FBCALL void         fb_DataReadByte     ( char *dst );
 FBCALL void         fb_DataReadUByte    ( unsigned char *dst );
 FBCALL void         fb_DataReadShort    ( short *dst );
