+++ src/rtlib/str_trim.c
@@ -2,15 +2,10 @@
 
 #include "fb.h"
 
-
-/*:::::*/
-FBCALL FBSTRING *fb_TRIM 
-	( 
-		FBSTRING *src 
-	)
+FBCALL FBSTRING *fb_TRIM( FBSTRING *src )
 {
 	FBSTRING *dst;
-	int len;
+	ssize_t len;
 	char *src_ptr;
 
 	if( src == NULL )
@@ -25,14 +20,14 @@ FBCALL FBSTRING *fb_TRIM
 		if( len > 0 )
 		{
 			src_ptr = fb_hStrSkipCharRev( src->data, len, 32 );
-			len = (int)(src_ptr - src->data) + 1;
+			len = (ssize_t)(src_ptr - src->data) + 1;
 		}
 	}
 
 	if( len > 0 )
 	{
 		src_ptr = fb_hStrSkipChar( src->data, FB_STRSIZE( src ), 32 );
-		len -= (int)(src_ptr - src->data);
+		len -= (ssize_t)(src_ptr - src->data);
 		if( len > 0 )
 		{
 			/* alloc temp string */
@@ -58,7 +53,3 @@ FBCALL FBSTRING *fb_TRIM
 
 	return dst;
 }
-
-
-
-
